Container_with_most_water.cpp
Problem Statement:  Given an integer array height of length n. There are n vertical lines drawn such that the two endpoints of the ith line are (i, 0) and (i, height[i]). Find two lines that together with the x-axis form a container, such that the container contains the most water. Return the maximum amount of water a container can store.

Notice that you may not slant the container.



Signed-off-by: MurthyJaya <116714949+MurthyJaya@users.noreply.github.com>