Vector_Erase.cpp
Problem Statement: You are provided with a vector of  integers. Then, you are given  queries. For the first query, you are provided with  integer, which denotes a position in the vector. The value at this position in the vector needs to be erased. The next query consists of  integers denoting a range of the positions in the vector. The elements which fall under that range should be removed. The second query is performed on the updated vector which we get after performing the first query.

Signed-off-by: Divyanshu Gupta <52619165+divyanshu799@users.noreply.github.com>