Segregate_even_&_odd_nodes_in_LinkedList.cpp
Problem Statement : Segregate even and odd nodes in LinkedList

Given a LinkedList of integers. Modify the LinkedList in such a way that in Modified LinkedList all the even numbers appear before all the odd numbers in LinkedList.

Also, note that the order of even and odd numbers should remain the same. 



Signed-off-by: Anuj Chaudhary <42280986+Anuj-Chaudhary@users.noreply.github.com>